refactor: view, controller -> presentation 레이어 분리
