+++ src/main/java/presentation/InputParser.java
@@ -1,4 +1,4 @@
-package controller;
+package presentation;
 
 import java.util.Arrays;
 import java.util.List;
