+++ src/main/java/presentation/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package presentation.view;
 
 import java.util.Scanner;
